Dishwasher remaining time UI & entrance light automation improvement
The dishwasher card appearing on the shortcut screen is now following the same design as the rest of the shortcut view page.

Entrance lights automations have also been improved to ensure, during the day, each side of the hallway works independently.
During the evening and night, both sides of the hallway stays turning on and off together.